requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pydantic-settings>=2.4
